+++ src/main/java/domain/exception/DuplicateUserNameException.java
@@ -1,0 +1,7 @@
+package domain.exception;
+
+public class DuplicateUserNameException extends RuntimeException {
+    public DuplicateUserNameException() {
+        super("이름 중복");
+    }
+}
