+++ src/test/java/domain/UserTest.java
@@ -1,13 +1,14 @@
 package domain;
 
+import domain.exception.InvalidNameException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class UserNameTest {
+class UserTest {
 
     @Test
     void 사람_이름_예외() {
-        assertThrows(InvalidNameException.class, () -> new UserName("exception"));
+        assertThrows(InvalidNameException.class, () -> new User("exception"));
     }
 }
